Fix account reset cancel race condition (#11329)
* add failing spec

* change database update strategy for account reset cancellation

* Do not send notifications if request is not successful

changelog: Bug Fixes, Account Reset, Fix race condition where concurrent requests to cancel could result in an exception

* absorb NotifyUserOfRequestCancellation into PendingRequestForUser

* Remove unused true/false return


* remove unnecessary allowed extra analytics

---------
