+++ spec/features/idv/phone_otp_rate_limiting_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.feature 'phone otp rate limiting', :js, allowed_extra_analytics: [:*] do
+RSpec.feature 'phone otp rate limiting', :js do
   include IdvStepHelper
 
   let(:user) { user_with_2fa }
